refactor(core): extract maintenance/init-vacuum.ts — initDatabase + vacuumDatabase
