+++ packages/core/src/maintenance.ts
@@ -1,9 +1,8 @@
-import { statSync } from "node:fs";
-import { assertSchemaReady, connect, getSchemaVersion, resolveDbPath } from "./db.js";
-import { withDb } from "./maintenance/with-db.js";
-import { ensureMaintenanceJobsSchema } from "./maintenance-jobs.js";
-import { bootstrapSchema } from "./schema-bootstrap.js";
-
+export { initDatabase, vacuumDatabase } from "./maintenance/init-vacuum.js";
+export {
+	compareMemoryRoleReports,
+	getMemoryRoleReport,
+} from "./maintenance/memory-role-report.js";
 export {
 	applyRawEventRelinkPlan,
 	applyRawEventRelinkPlanWithDb,
@@ -31,38 +30,6 @@ export type {
 	RawEventStatusItem,
 	RawEventStatusResult,
 } from "./maintenance/types.js";
-
-import { applyRawEventRelinkPlanWithDb } from "./maintenance/relink.js";
-
-export {
-	compareMemoryRoleReports,
-	getMemoryRoleReport,
-} from "./maintenance/memory-role-report.js";
-
-export function initDatabase(dbPath?: string): { path: string; sizeBytes: number } {
-	const resolvedPath = resolveDbPath(dbPath);
-	const db = connect(resolvedPath);
-	try {
-		if (getSchemaVersion(db) === 0) {
-			bootstrapSchema(db);
-		}
-		assertSchemaReady(db);
-		ensureMaintenanceJobsSchema(db);
-		applyRawEventRelinkPlanWithDb(db);
-		const stats = statSync(resolvedPath);
-		return { path: resolvedPath, sizeBytes: stats.size };
-	} finally {
-		db.close();
-	}
-}
-
-export function vacuumDatabase(dbPath?: string): { path: string; sizeBytes: number } {
-	return withDb(dbPath, (db, resolvedPath) => {
-		db.exec("VACUUM");
-		const stats = statSync(resolvedPath);
-		return { path: resolvedPath, sizeBytes: stats.size };
-	});
-}
 
 // ---------------------------------------------------------------------------
 // Reliability metrics
